+++ mindspore/ops/infer/ops_func_impl/masked_select.cc
@@ -50,8 +50,7 @@ BaseShapePtr MaskedSelectFuncImpl::InferShape(const PrimitivePtr &primitive,
   bool is_dynamic = IsDynamic(input_shape) || IsDynamic(mask_shape);
   if (!is_dynamic) {
     auto broadcast_shape = CalBroadCastShape(input_shape, mask_shape, primitive->name(), "input", "mask");
-    int64_t num = std::accumulate(broadcast_shape.begin(), broadcast_shape.end(), static_cast<int64_t>(1),
-                                  std::multiplies<int64_t>());
+    int64_t num = std::accumulate(broadcast_shape.begin(), broadcast_shape.end(), 1, std::multiplies<int64_t>());
     ShapeVector real_shape = {num};
     return std::make_shared<abstract::TensorShape>(real_shape);
   }
@@ -79,8 +78,7 @@ ShapeArray MaskedSelectFuncImpl::InferShape(const PrimitivePtr &primitive, const
   bool is_dynamic = IsDynamic(input_shape) || IsDynamic(mask_shape);
   if (!is_dynamic) {
     auto broadcast_shape = CalBroadCastShape(input_shape, mask_shape, primitive->name(), "input", "mask");
-    int64_t num = std::accumulate(broadcast_shape.begin(), broadcast_shape.end(), static_cast<int64_t>(1),
-                                  std::multiplies<int64_t>());
+    int64_t num = std::accumulate(broadcast_shape.begin(), broadcast_shape.end(), 1, std::multiplies<int64_t>());
     return {ShapeVector{num}};
   }
   return {input_shape};
