+++ mindspore/ops/kernel/ascend/aclnn/kernel_mod_impl/customize/group_norm_aclnn_kernel.cc
@@ -50,9 +50,9 @@ void GroupNormAscend::GetWorkSpaceInfo(const std::vector<KernelTensor *> &inputs
   const auto &x_shape = inputs[0]->GetShapeVector();
   n_ = x_shape[0];
   c_ = x_shape[1];
-  hw_ = (x_shape.size() == kNumberTwo) ? 1
-                                       : std::accumulate(x_shape.begin() + kIndex2, x_shape.end(),
-                                                         static_cast<int64_t>(1), std::multiplies<int64_t>());
+  hw_ = (x_shape.size() == kNumberTwo)
+          ? 1
+          : std::accumulate(x_shape.begin() + kIndex2, x_shape.end(), 1, std::multiplies<int64_t>());
 
   GetWorkspaceForResize(inputs[kIndex0], inputs[kIndex2], inputs[kIndex3], n_, c_, hw_, num_groups_, eps_,
                         outputs[kIndex0], outputs[kIndex1], outputs[kIndex2]);
