+++ mindspore/ops/kernel/ascend/aicpu/aicpu_ops/cpu_kernel/ms_kernel/fftnbase.cc
@@ -128,7 +128,7 @@ void FFTNBaseCpuKernel::FFTNGetInputs(CpuKernelContext &ctx) {
     calculate_shape_[dim_[i]] = s_[i];
   }
 
-  fft_nums_ = std::accumulate(s_.begin(), s_.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
+  fft_nums_ = std::accumulate(s_.begin(), s_.end(), 1, std::multiplies<int64_t>());
   norm_weight_ = GetNormalized(fft_nums_, norm_, forward_);
 }
