+++ mindspore/ops/kernel/cpu/native/adam_cpu_kernel.cc
@@ -172,8 +172,7 @@ int AdamCpuKernelMod::Resize(const std::vector<KernelTensor *> &inputs, const st
     return KRET_RESIZE_FAILED;
   }
 
-  input_elements_ =
-    IntToSize(std::accumulate(var_shape.begin(), var_shape.end(), static_cast<int64_t>(1), std::multiplies<int64_t>()));
+  input_elements_ = IntToSize(std::accumulate(var_shape.begin(), var_shape.end(), 1, std::multiplies<int64_t>()));
   input_elements_ = input_elements_ / LongToSize(batch_size_);
   if (batch_rank_ > 1) {
     if (var_shape.size() < lr_shape.size()) {
