+++ mindspore/ops/kernel/cpu/native/apply_ada_max_cpu_kernel.cc
@@ -100,12 +100,10 @@ int ApplyAdaMaxCpuKernelMod::Resize(const std::vector<KernelTensor *> &inputs,
   }
 
   if (!lr_shape.empty()) {
-    batch_size_ =
-      std::accumulate(lr_shape.begin(), lr_shape.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
+    batch_size_ = std::accumulate(lr_shape.begin(), lr_shape.end(), 1, std::multiplies<int64_t>());
   }
 
-  input_elements_ =
-    std::accumulate(var_shape.begin(), var_shape.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
+  input_elements_ = std::accumulate(var_shape.begin(), var_shape.end(), 1, std::multiplies<int64_t>());
   if (batch_size_ > 0) {
     input_elements_ = input_elements_ / batch_size_;
   } else {
