+++ mindspore/ops/kernel/cpu/native/dctn_cpu_kernel.cc
@@ -120,7 +120,7 @@ int DCTNCpuKernelMod::Resize(const std::vector<KernelTensor *> &inputs, const st
     calculate_shape_[dim_[i]] = s_[i];
   }
 
-  fft_nums_ = std::accumulate(s_.begin(), s_.end(), static_cast<int64_t>(1), std::multiplies<int64_t>());
+  fft_nums_ = std::accumulate(s_.begin(), s_.end(), 1, std::multiplies<int64_t>());
   int64_t cmpt_nums = fft_nums_ * static_cast<int64_t>(pow(kNormFactor, static_cast<int64_t>(s_.size())));
   norm_weight_ = GetNormalized(cmpt_nums, norm_, forward_);
   return KRET_OK;
